refactor to create rocket app in a module to be consumed by main
